[Complete] Get Activities By User.

+++ Batch4.FitnessTracker.Models/Db/Tbl_Activity.cs
@@ -15,9 +15,9 @@ namespace Batch4.FitnessTracker.Models.Db
         public int ActivityId { get; set; }
         public int UserId { get; set; }
         public int ActivityTypeId { get; set; }
-        public double Metric1 { get; set; }
-        public double Metric2 { get; set; }
-        public double Metric3 { get; set; }
-        public double CaloriesBurned { get; set; }
+        public decimal Metric1 { get; set; }
+        public decimal Metric2 { get; set; }
+        public decimal Metric3 { get; set; }
+        public decimal CaloriesBurned { get; set; }
     }
 }
