+++ Batch4.FitnessTracker.Models/Models/Activity/ActivityRequestModel.cs
@@ -10,7 +10,7 @@ public class ActivityRequestModel
 {
     public int UserId { get; set; }
     public int ActivityTypeId { get; set; }
-    public double Metric1 { get; set; }
-    public double Metric2 { get; set; }
-    public double Metric3 { get; set; }
+    public decimal Metric1 { get; set; }
+    public decimal Metric2 { get; set; }
+    public decimal Metric3 { get; set; }
 }
